+++ lib/src/builder/formy_selector.dart
@@ -3,12 +3,12 @@ part of 'form_manager.dart';
 abstract class FormySelector<Control, Value> extends StatefulWidget {
   const FormySelector({
     super.key,
-    required this.control,
+    required this.controller,
     required this.selector,
     required this.child,
   });
 
-  final Control control;
+  final Control controller;
   final Value Function(Control value) selector;
   final Widget Function(Value value) child;
 }
@@ -20,7 +20,7 @@ abstract class FormySelectorState<TValue, TSelected,
   @override
   void initState() {
     super.initState();
-    _value = widget.selector(widget.control);
+    _value = widget.selector(widget.controller);
     addListener();
   }
 
@@ -28,7 +28,7 @@ abstract class FormySelectorState<TValue, TSelected,
   void removeListener();
 
   void _onChanged() {
-    final newValue = widget.selector(widget.control);
+    final newValue = widget.selector(widget.controller);
     if (newValue != _value) {
       setState(() {
         _value = newValue;
@@ -53,7 +53,7 @@ abstract class FormySelectorState<TValue, TSelected,
 class GroupSelector<T> extends FormySelector<GroupController, T> {
   const GroupSelector({
     super.key,
-    required super.control,
+    required super.controller,
     required super.selector,
     required super.child,
   });
@@ -66,26 +66,26 @@ class _GroupSelectorState<T>
     extends FormySelectorState<GroupController, T, GroupSelector<T>> {
   @override
   void addListener() {
-    _FormManager._instance.insertGroup(widget.control);
-    widget.control.addListener(triggerUpdate);
+    _FormyFormManager._instance.insertGroup(widget.controller);
+    widget.controller.addListener(triggerUpdate);
   }
 
   @override
   void removeListener() {
-    widget.control.removeListener(triggerUpdate);
+    widget.controller.removeListener(triggerUpdate);
   }
 
   @override
   void dispose() {
     super.dispose();
-    _FormManager._instance.removeGroup(widget.control);
+    _FormyFormManager._instance.removeGroup(widget.controller);
   }
 }
 
 class FieldSelector<T> extends FormySelector<FieldController, T> {
   const FieldSelector({
     super.key,
-    required super.control,
+    required super.controller,
     required super.selector,
     required super.child,
   });
@@ -98,18 +98,18 @@ class _FieldSelectorState<T>
     extends FormySelectorState<FieldController, T, FieldSelector<T>> {
   @override
   void addListener() {
-    _FormManager._instance.insertField(widget.control);
-    widget.control.addListener(triggerUpdate);
+    _FormyFormManager._instance.insertField(widget.controller);
+    widget.controller.addListener(triggerUpdate);
   }
 
   @override
   void removeListener() {
-    widget.control.removeListener(triggerUpdate);
+    widget.controller.removeListener(triggerUpdate);
   }
 
   @override
   void dispose() {
     super.dispose();
-    _FormManager._instance.removeField(widget.control);
+    _FormyFormManager._instance.removeField(widget.controller);
   }
 }