+++ lib/src/widgets/formy_group_visibility.dart
@@ -3,7 +3,7 @@ import 'package:flutter_formy/flutter_formy.dart';
 
 class FormyGroupVisibility extends GroupSelector<bool> {
   FormyGroupVisibility(
-      {super.key, required super.control, required Widget child})
+      {super.key, required super.controller, required Widget child})
       : super(
           selector: (value) => value.state.isEnabled,
           child: (value) {